README: wire up the added hook to the code example
Allows to keep syntax highlight in markup modes and unsupported
languages.

I copied the code from my configuration because I have used it for ½ a
month, and can be sure it works right away. But if anybody request, it's
no problem for me to, instead, adapt the old code sample to the new
style.

Note: the example uses empty string instead of nil, which might cause
warnings in *Messages* buffer. However nil just doesn't work, an empty
string is the best way known so far to unset a face. See this for more
discussion: https://emacs.stackexchange.com/questions/35349/what-is-the-correct-way-to-unset-foreground-color

Signed-off-by: Constantine Kharlamov <Hi-Angel@yandex.ru>