color-identifiers-mode.el: don't run hooks manually (#76)
First of all, running the hook is automated by the `define-minor-mode`
macro. It creates the hook if it was not yet defined, and always automatically
runs it whenever the mode is enabled.

Second, the "else" branch where the

    (run-hooks 'color-identifiers-mode-hook))

code was located in my experiments was never ran. Not upon enabling the
mode, nor upon closing the buffer. I am not sure under what circumstance
it may run, tbh.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>