README.md: avoid warnings in face-remap example
Using it the way the example suggests will result in harmless warnings
in *Messages* buffer, described here¹.

    Unable to load color ""

As described on the same page, more correct way to drop unwanted
highlight would be to inherit the default style in face-remap call.

As a side note, I've been using this "default style remap" for 2 years
now with no problems.

1: https://emacs.stackexchange.com/questions/35349/what-is-the-correct-way-to-unset-foreground-color