Turn elisp-declarations-in-sexp to an interative function
First of, Emacs does not support tail-recursion optimization.  Second,
even if it did, the `elisp-declarations-in-sexp` could not be
tail-optimized, because it saves some stuff to an accumulator before
calling a recursion, and does not pass the accumulator further.

Using an iterative approach here should save lots of CPU cycles.

The correctness of the function was checked on a 10k lines file
`lisp/simple.el`, it returns the same identifiers before and after the
change.