Don't regenerate colors when emacsclient with a different UI appears
When GUI emacs is used and additionally terminal-based `emacsclient'
gets launched, that may cause colors for identifiers change globally.
This is not a good user experience. Ideally we should keep two
separate vectors, for graphics and terminal versions of colors. But
it's unclear how that should communicate with theme changes, so unless
there's demand, let's for now just make sure we don't toggle colors
back and forth everywhere.